GeometryCache: Fixed truncation warnings
#jira UE-173965
#rb JeanLuc.Corenthin

#changelist validated
#virtualized

[CL 31226032 by anousack kitisa in 5.4 branch]

+++ Engine/Plugins/FX/Niagara/Source/Niagara/Public/NiagaraRendererProperties.h
@@ -107,9 +107,9 @@ struct FNiagaraRendererVariableInfo
 
 	FNiagaraRendererVariableInfo() {}
 	FNiagaraRendererVariableInfo(int32 InDataOffset, int32 InGPUBufferOffset, int32 InNumComponents, bool bInUpload, bool bInHalfType)
-		: DatasetOffset(InDataOffset)
-		, GPUBufferOffset(InGPUBufferOffset)
-		, NumComponents(InNumComponents)
+		: DatasetOffset(static_cast<uint16>(InDataOffset))
+		, GPUBufferOffset(static_cast<uint16>(InGPUBufferOffset))
+		, NumComponents(static_cast<uint16>(InNumComponents))
 		, bUpload(bInUpload)
 		, bHalfType(bInHalfType)
 	{
